example/requirements.txt
grpcio
grpcio-tools
Pillow
numpy
googleapis-common-protos
